+++ app.js
@@ -14,15 +14,32 @@ var express = require('express');
 var bodyParser = require('body-parser');
 var auth = require("http-auth");
 var fs = require('fs');
-var logger = require('./helpers/logger');
+var cors = require('cors')
 var config = require('./config.js');
+var logger = require('./helpers/logger');
+var res_h = require('./helpers/response_handler');
 
 /********************************************/
 /* Config APP
 /********************************************/
+var current_version = "v1.2";
+
 port = process.env.PORT || config.port;
 
 var app = express();
+// Certs
+var options;
+if (config.https.toLowerCase() == "yes"){
+    options = {
+      key: fs.readFileSync(__dirname + '/ssl/server.key'),
+      cert: fs.readFileSync(__dirname + '/ssl/server.crt')
+    };
+}
+
+// CORS
+if (config.cors.toLowerCase() == "yes"){
+    app.use(cors());
+}
 
 // Basic authentication
 if (config.basic_auth.toLowerCase() == "yes"){
@@ -33,23 +50,56 @@ if (config.basic_auth.toLowerCase() == "yes"){
     app.use(auth.connect(auth_secure));
 }
 
-// Certs
-var options;
-if (config.https.toLowerCase() == "yes"){
-    options = {
-      key: fs.readFileSync(__dirname + '/ssl/server.key'),
-      cert: fs.readFileSync(__dirname + '/ssl/server.crt')
-    };
-}
 
 // Body
-app.use(bodyParser.json())
-app.use(bodyParser.urlencoded({extended: true}))
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({extended: true}));
+
+/**
+ * Versioning
+ * Using: Header: "api-version: vX.Y" or URL: /v1.2/
+ */
+app.use(function(req, res, next) {
+    var api_version_header = req.get('api-version');
+    var api_version_url = req.path.split('/')[1];
+    var regex_version = /^v\d+(?:\.\d+){0,1}$/i;
+    var new_url = "";
+
+    if (api_version_url && regex_version.test(api_version_url))
+        new_url = req.url;
+    else if (api_version_header && regex_version.test(api_version_header))
+        new_url = "/" + api_version_header + req.url;
+    else 
+        new_url = "/" + current_version + req.url;
+
+    req.url = new_url;
+
+    next();
+});
+
 
 // Controllers
-app.use(require('./controllers'))
+app.use("/" + current_version, require('./controllers'));
+//Example: app.use("/v1.2", require('./versions/v1.2/controllers'));
+
+
+// APP Errors
+app.use (function (err, req, res, next){
+    if ( err == "Error: invalid json" ){
+        logger.log(req.connection.remoteAddress + " " + req.method + " " + req.path);
+        res_h.bad_request("607", "", res);
+    }
+    else{
+        logger.log("Internal Error");
+        if(err.stack)
+            logger.log(err.stack);
+        logger.log("Exiting...");
+        process.exit(1);
+    }
+});
 
 /********************************************/
+
 
 // Create server
 if (config.https.toLowerCase() == "yes"){
@@ -65,3 +115,22 @@ else{
     });
 }
 
+
+// Event Handler
+process.on('uncaughtException', function(err) {
+    logger.log("Internal Error: uncaughtException");
+    if(err.stack)
+        logger.log(err.stack);
+    logger.log("Exiting...");
+    process.exit(1);
+});
+
+process.on('SIGTERM', function() {
+    logger.log("Exiting... (SIGTERM)");
+    process.exit(1);
+});
+
+process.on('SIGINT', function() {
+    logger.log("Exiting... (SIGINT)");
+    process.exit(1);
+});