+++ config.js
@@ -7,11 +7,12 @@ config.port = "55000";
 // Security
 config.https = "yes"; // Values: yes, no
 config.basic_auth = "yes"; // Values: yes, no
-config.AccessControlAllowOrigin = ["*"];
-config.AccessControlAllowHeaders = ["kbn-version"];
+config.BehindProxyServer = "no"; // Values: yes, no
+config.cors = "yes"; // Values: yes, no
 
 // Paths
 config.ossec_path = "/var/ossec";
+config.log_path = "/var/ossec/logs/api.log";
 config.api_path = __dirname;
 
 // Logs