+++ helpers/errors.js
@@ -28,10 +28,11 @@ errors['02'] = "Command output not in JSON";
 // 600 - 699 Requests
 errors['600'] = "Param not valid. Integer expected";  // Integer
 errors['601'] = "Param not valid. Valid characters: a-z, A-Z, 0-9, ., _, -,/,\\";  // Names
-errors['603'] = "No URL found. Bad HTTP verb?, typo?";  // Default error
+errors['603'] = "The requested URL was not found on this server";  // Default error
 errors['604'] = "Filter error";  // Filter
 errors['605'] = "Param not valid. Date format: YYYYMMDD";  // Date
 errors['606'] = "Param not valid. IP invalid";  // IP
+errors['607'] = "Invalid content-type. POST requests should be 'application/json' or 'application/x-www-form-urlencoded'";  // 
 
 exports.description = function(n){
     if (n in errors)