+++ scripts/check_config.py
@@ -10,8 +10,8 @@ import os
 if __name__ == "__main__":
     ossec_path = "/var/ossec"
     r_error = 0
-    r_description = ""
-    r_response = ""
+    r_message = ""
+    r_data = ""
     
     output = ""
     err = ""
@@ -30,22 +30,22 @@ if __name__ == "__main__":
         if err:
             if "Error" in err:
                 r_error = 82
-                r_description = "{0}".format(lines[error_line:-1])
+                r_message = "{0}".format(lines[error_line:-1])
             else:
                 r_error = 0
-                r_response = "OK"
+                r_data = "OK"
         else:
             r_error = 81
-            r_description = "Error unknown."
+            r_message = "Error unknown."
     except Exception as e:
         r_error = 80
-        r_description = "Problem running command: {0}".format(e)
+        r_message = "Problem running command: {0}".format(e)
 
     # Response
     response = {'error': r_error}
     if r_error == 0:
-        response['response'] = r_response
+        response['data'] = r_data
     else:
-        response['description'] = r_description
+        response['message'] = r_message
 
     print(json.dumps(response))